+++ console.html
@@ -6,6 +6,6 @@
   <title>Document</title>
 </head>
 <body>
-  <script src="js/main.js"></script>
+  <script src="js/main.js" type="module"></script>
 </body>
 </html>
