+++ js/dialog.js
@@ -1,0 +1,9 @@
+// const avatar = ['01', '02', '03', '04', '05', '06', '07', '08'];
+// const TYPE = ['palace', 'flat', 'house', 'bungalow'];
+// const TIME = ['12:00', '13:00', '14:00'];
+// const FEATURES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+// const PHOTOS = ['http://o0.github.io/assets/images/tokyo/hotel1.jpg', 'http://o0.github.io/assets/images/tokyo/hotel2.jpg', 'http://o0.github.io/assets/images/tokyo/hotel3.jpg'];
+// const TITLE = ['Вашему вниманию представляю квартиру!', 'Отличные апартаменты в самом центре!', 'Уютная квартира неподалёку от парка с детской площадкой'];
+// const similarObjectsNear = 10;
+
+// export {avatar, TYPE, TIME, FEATURES, PHOTOS, TITLE, similarObjectsNear};
